+++ include/lug/System/Memory/Arena.hpp
@@ -22,7 +22,7 @@ public:
 
     ~Arena() = default;
 
-    void* allocate(size_t size, size_t alignment, const char* file, int line);
+    void* allocate(size_t size, size_t alignment, const char* file, size_t line);
     void free(void* ptr);
     void reset();
